sdk/state: add fully signed param to confirm functions (#76)
Add fully signed param to confirm functions that indicates if the value being returned is completely signed and valid for submission if the user wished to submit the transactions to the network. Also, the error value of the confirm functions no longer error when the value is not completely signed, and no other error occurred.

### Why
@acharb mentioned it was confusing that the functions returns values that would be usable when returning an errors. I also found it difficult to remember that the errors returned from these functions could be catastrophic, but then also just routine part of how the process works.

It's a bit more verbose but clearer if the function returns some solid signal that this is fully signed, or complete and submittable, separate to whether some other arbitrary error occurred.

This seemed like a good use case for error inspection, but in reality it was more confusing.