sdk/state: keep amounts in int64 (#113)
Change the asset limit to be an int64 to be consistent with the Amount Balance being stored as an int64.

For consistency, so all amounts are stored the same way. This also makes it easier to work with the value since we can use simple math operations on it if we need to.