sdk/state: consolidate the building of close txs (#131)
Consolidate the building of close transactions so that there is only one function and one place that occurs.

Building the close transactions happens in three places currently, in CloseTxs, PaymentTxs, and in ProposePayment. Each location collect the input data in different ways. There should be one owner responsible for close tx building. Ideally we are always building close txs during the update or final steps the exact same way.