sdk/state: remove CoordinatedCloseTxs and build observation into closing agreements (#132)
Remove `CoordinatedCloseTxs` and build the observation parameters into the `ClosingAgreement`.

Coordinated close is the happy path and so we should make it the default way to close, with closing without the additional signatures being the fallback. There's a few changes we need to make for this to happen, and this is the first, that there aren't separate functions, like `CoordinatedCloseTxs` for getting the coordinated closing transaction. The user should be able to just get the closing transactions, via `CloseTxs`, and if the agreement is a coordinated close agreement, it should just work.

There was another reason we were thinking of moving observation parameters into the closing agreement, but I don't remember what they were.