sdk/state: remove Coordinated term from func names (#136)
Remove the term `Coordinated` from function names and code.

The default happy path is to do a coordinated close, or rather, it should just be part of the standard close process which involves declaring, trying to coordinate the immediate close, and falling back to using the existing close transactions. While we call this coordinated close in the SEP, I don't think we get much value of using the term in the SDK because in practice there will be one close process not two.