sdk/state: fix coordinated close agreement declaration signatures (#134)
When constructing the coordinated close agreement include declaration signatures of the original close agreement.

The coordinated `CloseAgreement` doesn't contain declaration signatures. It has a slice for them but it is empty. The coordinated close process replaces the current close agreement, and so for users that choose to coordinate before closing, they will lose signatures for the declaration transaction when the new agreement is promoted to authorized.