sdk/state: fix coordinated close for confirming participant (#135)
Build the final authorized close agreement for a confirmed coordinated close off the last authorized rather than last unauthorized.

The confirmer can't execute the coordinated close because it has the signatures from the unauthorized transaction added in, but the unauthorized agreement might be a payment that was abandoned and for the confirmer it definitely isn't the close that is being amended. For the confirmer of the coordinated close the agreement should match their authorized, we check that at the beginning, the only thing that is different is the close signatures so we can base it off the rest and then adopt the signatures if they are valid.

This is a little odd in that the proposer sets it as their unauthorized, but that doesn't really matter because they also have the same agreement, except for the close signatures, set in their authorized so nothing is lost.