sdk/state: reject payments from remote if they change balance in their favor (#138)
Change the `ConfirmPayment` function so it rejects payments from the remote if the remote is proposing that the local pay the remote.

This is a payment channel where either party can send money to the other party, and it isn't designed fro requesting/invoicing the other party for funds. There's several reasons for this. It is safer and easier to write confirm functionality that enforces this rather than require that the user always check the balance is an okay change. The channel is designed to always move forward and in disagreement it becomes unclear if the channel should close or stay open and skip a generation. Mostly this is for simplicity for the major use case.