sdk: enable formatting checks (#140)
Enable formatting checks in CI with `golangci-lint`.

I noticed when editing code that there's some code that isn't formatted with common Go formatting tools `go fmt`/`goimports`/`gofumpt`. `golangci-lint` supports format checking and a lot of linters, but it only has a small subset turned on by default and that doesn't include any format checks.

In terms of why `gofumpt`: `gofmt` does formatting and simplification. `goimports` is a superset of `gofmt` without simplification, but keeps import directives up-to-date and is conventionally almost always preferred over `gofmt` now for everything except simplification. `gofumpt` is a superset of `goimports` and is also common in some IDEs like VSCode and makes code more the same, so more code looks the same.

Like the other linters these checks can be run locally by simply running `golangci-lint run`, and formatting can be updated with `golangci-lint run --fix`.

This is somewhat unrelated to this PR, but there's one linter already enabled, govet, that we should enable an optional feature for, the shadow checks. I think we should enable it because stellar/go uses it and if we move this code into stellar/go it will be easier if our code is passing the same linters and is at an equal or higher standard. I have commented it out because it doesn't pass at the moment and I don't want to spend the time fixing that atm, but will at some point when I have a moment.