sdk/state: replace txnbuild.Asset with an internal simpler asset type (#149)
Replace txnbuild.Asset in the state package with an internal simpler asset type, a string that follows [SEP-11 Txrep]'s format.

While experimenting with the state package (#122) I found the interface type txnbuild.Asset inconvenient. Interfaces can't be used to deserialize. While working around that issue I realized we don't really need an interface for what we're doing, we only need a simple value and it is more convenient to be able to specify a concrete value in the simplest form and to convert it to what the txnbuild package needs at the point it needs it.

I chose a string because we have a specification for assets as string in [SEP-11 Txrep] which has been used in Horizon for new endpoints that reference assets.

[SEP-11 Txrep]: https://stellar.org/protocol/sep-11