sdk/state: make formation expiry deterministic (#153)
Make the formation transaction that gets built deterministic by changing the time bounds to be based on time.Now to be based on a fixed time the proposer chooses.

The formation transaction sets its time bounds relative to the now time, which requires both participants to have synchronized clocks at a resolution of 1 second and for the clients to generate the transaction at the exact moment. This is unlikely to happen. This behavior also makes it impossible to regenerate the formation transaction later.