sdk/state: remove assets from places only needed for multi asset support (#156)
Remove assets from `CloseAgreements` and some other places.

In #72 we rewound support for multiple assets and moved forward with single asset support only. `CloseAgreements` only need to include asset information if the channel is multi-asset. This simplifies several things. 

I realized this could be simplified when I worked on #155. In that change I needed to add a balance to `EscrowAccount`. It was going to be significantly harder to follow the existing pattern and include asset information in that object, so I didn't. I think it is worth bringing the same simplicity to close agreements and everywhere else.