sdk/state: validate balances and handle deposits and balance changes (#155)
When receiving a proposed new close agreement/payment, error with `ErrUnderfunded` if the close agreement over commits the payer. Support updating the known balances of escrow accounts so the payment can be confirmed if on network the balance has been increased.

The state machine does not validate that a proposed close agreement will be satisfied by the balance of the paying participant's balance. This is a problem because the payer could over commit, breaking the channel. The receiver in a payment shouldn't accept a close agreement if it can't be satisfied by the current balances in the escrow accounts.

It's possible for participants to deposit more of an asset into their escrow account, so it should also be possible 
to update the known balances inside the channel when that happens. Deposits can happen at anytime without a participants involvement so the flow must support a way to move forward if a participant proposes a payment that based on stale knowledge is invalid, but would be valid with updated knowledge. The ConfirmPayment function returns an `ErrUnderfunded` error so that a caller can respond to that specific situation by checking the balance of the payers escrow account on network, updating it locally, and trying again.