sdk/state: add some helpful funcs (#177)
### What
Add some functions that I've found helpful:
 - Equal functions on agreements.
 - Functions to get the escrow account objects from the channel.

### Why
It's helpful for the user to be able to compare agreements to know if an agreement they have is the same as a new agreement.

It's helpful to be able to access the full state of the escrow account as known by the channel, and not just the balance that we expose now.

### Known Limitations
I shortcutted writing the equals methods by using the cmp.Equal package. We should really write a proper equals package that doesn't use reflection under the surface, but this should be sufficient for now.