sdk/state: refactor the open/close tx funcs (#181)
Refactor the open and close tx functions so that they are no longer exported, focus solely on the txs they needs to generate, and add new open and close tx functions that are exported and intended to be used by a user wanting to submit those transactions.

While working on #122 it became clear that the example code had to know how to get an agreement and a transaction and get the transaction into a state that was ready for submission. This is something the state machine can really take ownership of. I think this improves the API.