all: update quickstart to include CAP-21 extraSigners and CAP-40 (#192)
Update the quickstart image to include CAP-21 extraSigners and CAP-40. Also added in additional retry mechanics.

#189 and #190 introduce use of CAP-21 extraSigners and CAP-40, but I want to introduce the upgrade of the quickstart image and the new XDR independent of using the new functionality.

The change includes additional retry mechanics because after updating Horizon and Core, which included bringing in 2 months of changes from both respective main/master branches, there is an odd condition occurring when trying to submit a regular payment transaction during test setup. The payment occurs immediately after creating the new account and for some reason Horizon sees the transaction get a bad sequence number result from Core, but discards the response. Immediately resubmitting the transaction is successful. It's going to take a little while to nut this problem out and to understand if this is a problem in Core or Horizon, and if this is a problem in code that exists on the master branch of either or in our forks. The retry is a hold over so that we can keep working an building on CAP-40 without being blocked on this issue.