sdk/state: refactor signatures (#209)
Refactor how signatures are stored:
- Store the raw signature and not a decorated signature, not storing the hint.
- Store signatures in fields named for their purpose, rather than in a slice.
- Verify and sign in a single block, rather than piece meal.

Signatures were stored as decorated signature lists because early on it made it simple to add those signatures to a transaction at a later point. That optimization has made it painful to make several changes over the last few weeks, the worst being the implementation of CAP-40 (#190, #189).

The only benefit we got from storing decorated signatures was to speed up looking up signatures in each slice, because we frequently need to only verify a single signature for a single signer. Besides the fact it is unnecessary for us to be optimizing things like that this early in the project, we don't need the signatures stored in a slice, and therefore don't need the optimization. Storing signatures in a field specific for its use simplifies some of our code because we no longer need to go searching for specific signatures.

Decorated signatures are also inconvenient because in reality we can't just append them as they are anymore. Depending on where the signature is being used we must construct decorated signatures differently. This is because the hint of a payload signer is different to that of a regular signer, even for the same signature.

We also won't use decorated signatures when we implement recovery from disclosed signatures on Horizon, because Horizon doesn't public the decorated portion.

The change to sign in a single block is not particularly meaningful, it is just easier to write the signing and verification that way during the refactor.

Unfortunately this change is fundamental, touching many components, and I could not find an elegant way to isolate portions of the change that would be practical in isolation. I tried removing hints without. Thankfully many of the tests, including integration tests, did not need significant changes as they operate at a higher level, which provides some confidence that the change is sound.

Co-authored-by: Marcelo Salloum dos Santos <marcelo@stellar.org>