sep: add discussion about risks with queueing payments (#250)
Add discussion about the risks of queueing multiple payments by sending multiple payments to the other participant before they confirm each individually.

CAP-40 gives us a bonus benefit that participants can make multiple payments without waiting for the confirmation from teh receiver. Some future version of the protocol defined in the SEP could use that bonus and make it possible for participants to speed up and slow down independent of each other and be allowed to fall behind for a period or indefinitely. There is one side-effect and tradeoff to that though which is important to weigh with any implementation that uses it. I wanted to document it so that we didn't forget about it if we went all in on that approach in the future.