sdk/agent: repair failing payment due to stale balance data (#254)
Update escrow account on-network balance that is cached in the channel whenever sending or receiving a payment that looks like it exceeds the last known available balance.

If a payment is exceeding the balance the only option we have is to stop and close the channel, so we may as check the network to see if the balance is updated which might allow the payment to succeed.

I think the usefulness of this balance update is questionable given we will soon have the agent using the state's ingestion to update the balances, but in the interim I need to update the balances one way or another to get tests I'm writing for #225 and #226 to pass. This seems like a good temporary hold over and something we can evaluate in the near future if we still need.

I removed the TODO to remove the existing check because it's a little fuzzy to me if we really want to do that.

There's no test in this PR for this change because I'm writing tests for the agent in #242 which incorporates a few things that make writing tests for the agent much more straight forward.