sep: add security notes about trustline liabilities (#252)
Add notes to the security sub-sections about trust line liabilities.

It is important when we lookup the available balance of a trust line that we take into account that there could be offers setup on the account that reduce the available balance. The amount that a balance can be reduced by standing offers is the selling liabilities.

It's also worth mentioning buying liabilities where trust line limits are discussed, however this has no bearing on implementations because the possibility of exceeding trust line is an unsolved problem in general and only a problem for assets with excessive supply.