sdk/agent: add a first happy path test (#255)
Add a first happy path test that exercises two agents that are connected to each other and successfully open, pay, and close.

The agent code was ported over from the console example. The console example was really written to be throw away code, but the agent code within it was well formed enough to be something to build on. It's important we add tests early for the agent before we build too much more into it. This is a first step to closing #225.