sdk/agent: implement event handling with channels (#257)
Add event handlers for the things that can occur in the background during open, payment, and close.

The agent receives messages from the other participant containing confirmations/responses to requests made by this recipient, and requests from the other participant. Some of these operations that take place in the background need to be made known to the application using the agent.

The agent will also monitor the state of the network for changes that require action, and once that logic is added we'll need to add events using the same pattern.

A channel is used for delivering these background events because it makes it easier for developers who don't want to handle concurrent routines in their code. They can simply use a buffered channel and poll when they want to. Or they can put it in a go routine and trigger any functions they have for each.

This change is somewhat limited because some of the change depends on ingestion having been integrated into the agent, but I'm deferring doing that integration until the integration and state changes have been finished in the state package. Because of that there are some TODOs, but I think it is worth merging this and revisiting it during the ingestion work.