sdk/state: ingest from raw tx xdr and support fee bump txs (#260)
Ingest raw tx XDR instead of objects into the state ingest functions, and add support for fee bump txs as input.

When I started hooking the state machine's `IngestTx` function into the agent I discovered I was going to have to do this somewhere, and it seems like the most natural thing is to make the ingestion functions handle fee bumps. The easiest way to accept either a fee bump or non-fee bump tx is to accept the XDR and let the function figure it out. XDR is easier than using an object because the `txnbuild.GenericTransaction` type that wraps fee bumps and normal transactions is not constructible from outside the txnbuild package which means we'd have to convert to XDR first then create a `txnbuild.GenericTransaction` with that XDR, then pass it in. We may as well do that inside the `IngestTx`. This isn't a huge shift in philosophy because the `IngestTx` function already accepts other information about the transaction as XDR, so if anything it makes it more consistent.