sdk/agent: address data races (#262)
Add a mutex to the `agent.Agent` that serializes any reads and writes to state that is not immutable, namely the internal `state.Channel` and the other variables that record the other participants escrow details.

There is a data race in the agent because the agent runs a go routine in the background that receives messages from the other participant. It is possible for a message to arrive inbound while a payment is being proposed or other operation is being started.

I chose a mutex over using channels to interact with the `state.Channel` object because it seems like the more suitable tool. We could use a channel to pass commands to a background routine that operated on the channel, or we could even pass functions to a background routine to serialize them. Both of these alternative approaches using channels would be more complicated because data would also need to be passed back and so the sending routine needs to block on a returning value. Because of this I think a mutex is simpler and easier to understand the implications.

Not all functions in the agent need to be serialized. Some functions do not utilize any of the mutable fields of the agent.

Functions that access only the `conn io.ReadWriter` may not necessarily need locking because that field will typically be a `net.Conn` and `net.Conn`s are safe to use concurrently because their reads are concurrent and their writes use an internal lock. However, I have chosen to treat any function using the `conn io.ReadWriter` as requiring a lock since there is no guarantee the value is a `net.Conn` since the type of the field is `io.ReadWriter`. I don't think this will have any material impact.