sdk/agent: hook horizon payment stream into state ingestion (#265)
Hook horizon payment stream into state ingestion by streaming all transactions that occur on the network, piping them through to the state machine, and then checking if the state changes and broadcasting events to the application if the state changes.

The state package supports ingesting transactions that occur on network. The agent is responsible for providing those transactions to the state package. The horizon payment stream seems like the most straight forward source for transactions on the escrow account. The agent also sends events back to the application, and at the moment it does it in a rather incomplete fashion based on changes in the state of the state machine, however some events such as the open and close are more accurately triggered when those transactions appear on network.

The horizon integration is somewhat imperfect in that it ingests every transaction seen on network, not only transactions that occur for the escrow accounts. It works this way because horizon only supports streaming a single accounts transactions or all transactions. It might be possible to ingest transactions for two accounts separately but that introduces new complexity because transactions may be duplicate across both accounts or may be processed out-of-order. I've opened a feature request on horizon, stellar/go#3874, that would improve this, but right now this is the best we can do. (Note that I also considered using the Go SDK's `ingest` package to setup custom ingestion from history archives, but that complicates our test plans as it adds yet another component that we're dependent on. It also suffers the same problem as the current implementation in that it will ingest all transactions anyway, so offers no advantage over Horizon.)