sdk/agent: add NewAgent and Config (#272)
Add `NewAgent` and `Config` to the `sdk/agent` package and use those to setup the `agent.Agent`.

That's the same pattern we used in the `sdk/state` package for the `state.Channel`. It seemed to work pretty well there. It makes it really clear what are things that get set at start. It's helpful if we use the same pattern.