sdk/state: reorder funcs (#276)
Move the `Config` and `NewChannel` type and function above the `Channel` type.

The type and function are the entry point to using the `Channel`, but they're buried below some other types and functions. Moving them to the top makes them easier to discover and matches how the same pattern in the `sdk/agent` package.