sdk/state: store starting sequence in open agreement (#278)
Store the starting sequence in the open agreement.

The starting sequence is an assumed value that implicitly forms part of the agreement, however it would be better if it was an explicit part of the agreement and stored there rather than store inside the channel as a global value.

As a bonus this lets an escrow account holder to state that the open will occur at some future sequence number. On the surface this might be a cause for concern for the responder but it doesn't really change the security of anything since the initiator could always shift their sequence number after starting the open handshake. This just provides some flexibility and improves the debug-ability in cases that the two participants for some reason think the starting sequence is different.

I didn't plan on doing this now except that was I work on #227 (#273) it is inconvenient that the escrow accounts have their sequence numbers stored and what is stored there is important, so that led me to moving this up.