sdk/agent, sdk/state: add persistence (#273)
Add methods to `state.Channel` for snapshotting and restoring a snapshot. Add a callback to `agent.Agent` for snapshotting the state of the agent and its channel, and a function for restoring that state.

It's necessary for the channel state to be persisted somewhere reliably so that an application using them can be stopped and restarted.

This approach of calling back with a snapshot of the state defers any need for the agent to choose a persistence method or define what persistence might look like. It could be more performant to have the agent write directly to a database, but there isn't a huge amount of data to store since the agent doesn't need to store data for every transactions processed. For where this project is write now storing the state in a blob seems like the shortest path to experimentation.