sdk/horizon: split up horizon type (#288)
Move the functionality in the `horizon.Horizon` type into `horizon.BalanceCollector`, `horizon.SequenceNumberCollector`, `horizon.Submitter`, and `horizon.Streamer`.

For #275 I'm adding some fields to the type that holds the Horizon `Streamer` implementation and it is a bit odd to have these fields on the `horizon.Horizon` type when those fields only relate to a subset of that types functionality. It's a bit of an anti-pattern that the `horizon.Horizon` type holds a bunch of unrelated functionality. Splitting it up helps the change I'm going to be making clearer.

All the logic is a copy and paste in this change, no logic has been changed.