sdk/agent: handle connection loss (#297)
Exit the agent's receive loop early if EOF reached on agent's connection.

If an EOF error is seen there's no point attempting to read the connection again as that is an indication the connection is closed. In the example app this is clear because when one side disconnects the agent continuously logs the EOF error.