sdk/state: store and reuse txs (#294)
Store transactions that the state machine builds, and reuse them, rather than rebuilding as they are needed.

Our initial plan was to rebuild transactions as needed, however if the Stellar Go SDK that we use to build txs ever changes the transactions being built even in the slightest the signatures stored won't match the rebuilt transactions. Storing transactions at the point in time we generate the signatures is the only way to ensure we can submit that same transaction in the far future.

This change also stores the transaction hashes alongside them, since it is small effort to do so. Storing both the transactions and their hashes has the bonus that it makes debugging easier since the snapshotted state of the channel includes its transactions.